"Game of two Stacks" problem C++stack classes (Implemented with dynamic arrays)
I created an Stack class with some few operations. Basically I transfer as many elements I can on any of those stacks, then I Do the same with the other (and keep comparing the amount of elements with the last one) , If There's no possibility I push back the last element I got of the first Stack and keep trying.