Added Temeperature Convereter Project in CPP
In this project, I have build a simple temeperature converter in C++.
First it will ask the user to choose the type of conversion.
Then temperature will prompted from user and program will convert it in required unit.
The user have the option to do more than one conversion and can exit anytime.